refactor: remove `FloatType` and all of its usage (#74)
* refactor: remove FloatType from cmap2::tests

* refactor: remove FloatType from generation tests

* refactor: rewrite vector tests to remove FloatType & use macros

* refactor: add more macros to vector tests

* refactor: remove FloatType from vertex tests

* refactor: rewrite coords test using macros

* refactor: delete FloatType

+++ honeycomb-core/src/cells/orbits.rs
@@ -219,10 +219,10 @@ impl<'a, T: CoordsFloat> Iterator for Orbit2<'a, T> {
 
 #[cfg(test)]
 mod tests {
-    use crate::{common::FloatType, CMap2, DartIdentifier, Orbit2, OrbitPolicy};
+    use crate::{CMap2, DartIdentifier, Orbit2, OrbitPolicy};
 
-    fn simple_map() -> CMap2<FloatType> {
-        let mut map: CMap2<FloatType> = CMap2::new(6);
+    fn simple_map() -> CMap2<f64> {
+        let mut map: CMap2<f64> = CMap2::new(6);
         map.one_link(1, 2);
         map.one_link(2, 3);
         map.one_link(3, 1);
